Create hooks to control before and after actions
As a basic idea, hooks are used to manipulate the data before of after
it is added or requested to be added to the database.  I did not as yet
completely understand the mechanism of configuring hooks but seems to
be able to grasp it shortly.  Prepared custom hooks and built in hooks
as in tutorial.